Clean up NixOSContext setup
